requirements.txt
torch>=1.11.0
torchvision>=0.12.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
tqdm>=4.62.0
PyYAML>=5.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
pathlib
argparse